+++ blackmamba/script/refactoring/__init__.py
@@ -1,0 +1,18 @@
+"""
+All refactoring scripts are limited to open single file only.
+
+Rename
+------
+
+.. automodule:: blackmamba.script.refactoring.rename
+
+Expand star imports
+-------------------
+
+.. automodule:: blackmamba.script.refactoring.expand_star_imports
+
+Organize imports
+----------------
+
+.. automodule:: blackmamba.script.refactoring.organize_imports
+"""